+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Clean up error handling
+- Add custom launcher commands for a `seed`
 
 ## [0.1.0] - 2025-09-10